Refactor: Rename 'require-image-overlayed' to 'require-image-optimized'
- Update rule name from 'require-image-overlayed' to 'require-image-optimized'
- Update component path to 'src/components/image/ImageOptimized'
- Modify test cases and ESLint configuration to reflect new rule and component name
